fix(font): correct MSDF bilinear sampling with half-pixel offset
GPU texture sampling treats coordinates as texel centers (at 0.5, 1.5, etc.).
Our bilinear interpolation was treating them as edges, causing vertical
banding artifacts in MSDF font rendering.

Changes:
- Subtract 0.5 before computing integer texel coords to match GPU behavior
- Fix median3f formula (was returning wrong value for some inputs)
- Add sh_font_sample_msdf_nearest() for debugging (MSDF_USE_NEAREST flag)
- Improve font_compare.c with buffer verification for Z glyph

The half-pixel offset fix ensures CPU rendering matches WebGL output.

+++ shared/include/sh_font.h
@@ -260,7 +260,7 @@ typedef struct {
  */
 void sh_font_render_glyph(uint8_t *pixels, int buf_width, int buf_height,
                           const SHFont *font, const SHGlyph *glyph,
-                          int x, int y, float font_size,
+                          float x, float y, float font_size,
                           uint8_t r, uint8_t g, uint8_t b, uint8_t alpha,
                           float threshold);
 
@@ -274,7 +274,7 @@ void sh_font_render_glyph(uint8_t *pixels, int buf_width, int buf_height,
  */
 void sh_font_render_glyph_clipped(uint8_t *pixels, int buf_width, int buf_height,
                                    const SHFont *font, const SHGlyph *glyph,
-                                   int x, int y, float font_size,
+                                   float x, float y, float font_size,
                                    uint8_t r, uint8_t g, uint8_t b, uint8_t alpha,
                                    float threshold, const SHScissor *scissor);
